Fix backup bug, indicate some features not implemented, add settings.py
Backing up pokemon multiple times doesn't work because PKs need to
be unique and there's not "update an existing record" functionality.
Added a message that gracefully handles issue (without crashing) and
tells users about this missing feature.

Search feature is not implemented, so I now handle that a little more
gracefully.

Settings.py is added with the SECRET_KEY removed.